peer-manager: split dynamic neighbor helpers
