+++ src/peer_manager.rs
@@ -32,8 +32,11 @@ use crate::policy_admin::{
     neighbor_set_references, peer_group_references, policy_references,
 };
 
+mod dynamic;
 mod events;
 mod snapshot;
+
+use dynamic::{DeadLetteredPending, DynamicRange};
 
 const DEFAULT_HOLD_TIME: u16 = 90;
 const DEFAULT_CONNECT_RETRY_SECS: u32 = 5;
@@ -128,32 +131,6 @@ struct ManagedPeer {
 struct PendingInbound {
     handle: PeerHandle,
     session_id: u64,
-}
-
-/// Resolved dynamic neighbor range used for prefix matching at connection time.
-struct DynamicRange {
-    addr: std::net::IpAddr,
-    prefix_len: u8,
-    peer_group: String,
-    remote_asn: u32,
-    description: Option<String>,
-}
-
-/// Snapshot of a removed dynamic peer's unfired hot-apply intent. Carried
-/// across the auto-removal that fires when a dynamic peer goes back to
-/// Idle so a re-establishing peer at the same address inherits the retry.
-/// Without this, a transient TCP drop on a `[[dynamic_neighbors]]` peer
-/// that was mid-`pending_refresh` would silently drop the unfired Route
-/// Refresh / export-apply — same correctness risk that
-/// `ManagedPeer::pending_refresh` / `pending_export_apply` exist to close.
-/// The RFC 8326 initiator toggle lives here too: dynamic auto-removal drops
-/// the whole `ManagedPeer`, so this side table is the only place to preserve
-/// an operator's maintenance-window `GShut` toggle across re-establishment.
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
-struct DeadLetteredPending {
-    refresh: bool,
-    export_apply: bool,
-    graceful_shutdown: bool,
 }
 
 /// Manages the lifecycle of all peer sessions.
@@ -324,25 +301,6 @@ impl PeerManager {
         // remains visibly outside the peer-manager allocated range.
         self.next_session_id = self.next_session_id.wrapping_add(1).max(1);
         id
-    }
-
-    fn parse_dynamic_ranges(config: &Config) -> Vec<DynamicRange> {
-        config
-            .dynamic_neighbors
-            .iter()
-            .filter_map(|dn| {
-                let parts: Vec<&str> = dn.prefix.split('/').collect();
-                let addr: std::net::IpAddr = parts.first()?.parse().ok()?;
-                let prefix_len: u8 = parts.get(1)?.parse().ok()?;
-                Some(DynamicRange {
-                    addr,
-                    prefix_len,
-                    peer_group: dn.peer_group.clone(),
-                    remote_asn: dn.remote_asn,
-                    description: dn.description.clone(),
-                })
-            })
-            .collect()
     }
 
     fn build_transport_config(&self, config: &PeerManagerNeighborConfig) -> TransportConfig {
@@ -1379,37 +1337,6 @@ impl PeerManager {
         Ok(())
     }
 
-    /// Check whether a peer IP falls within any configured dynamic neighbor range.
-    fn match_dynamic_range(&self, addr: IpAddr) -> Option<&DynamicRange> {
-        self.dynamic_ranges.iter().find(|r| {
-            match (addr, r.addr) {
-                (IpAddr::V4(peer), IpAddr::V4(net)) => {
-                    if r.prefix_len > 32 {
-                        return false;
-                    }
-                    let mask = if r.prefix_len == 0 {
-                        0u32
-                    } else {
-                        u32::MAX << (32 - r.prefix_len)
-                    };
-                    (u32::from(peer) & mask) == (u32::from(net) & mask)
-                }
-                (IpAddr::V6(peer), IpAddr::V6(net)) => {
-                    if r.prefix_len > 128 {
-                        return false;
-                    }
-                    let mask = if r.prefix_len == 0 {
-                        0u128
-                    } else {
-                        u128::MAX << (128 - r.prefix_len)
-                    };
-                    (u128::from(peer) & mask) == (u128::from(net) & mask)
-                }
-                _ => false, // IPv4/IPv6 mismatch
-            }
-        })
-    }
-
     async fn spawn_pending_inbound(&mut self, peer_addr: IpAddr, stream: TcpStream) -> bool {
         if self
             .peers
@@ -1647,66 +1574,6 @@ impl PeerManager {
                 }
             }
         }
-    }
-
-    /// Snapshot any unfired hot-apply / Route Refresh intent for a peer
-    /// about to be auto-removed, so a re-establishing peer at the same
-    /// address inherits the retry. No-op if neither flag is set.
-    /// Bounded at `dynamic_neighbor_limit` — over-cap evicts an
-    /// arbitrary entry with `warn!` to surface pathological churn.
-    fn dead_letter_pending_for(&mut self, peer_addr: IpAddr) {
-        let Some(managed) = self.peers.get(&peer_addr) else {
-            return;
-        };
-        if !managed.pending_refresh
-            && !managed.pending_export_apply
-            && !managed.advertise_graceful_shutdown
-        {
-            return;
-        }
-        let entry = DeadLetteredPending {
-            refresh: managed.pending_refresh,
-            export_apply: managed.pending_export_apply,
-            graceful_shutdown: managed.advertise_graceful_shutdown,
-        };
-        let cap = self.dynamic_neighbor_limit as usize;
-        if cap > 0
-            && !self.dead_lettered_pending.contains_key(&peer_addr)
-            && self.dead_lettered_pending.len() >= cap
-            && let Some(victim) = self.dead_lettered_pending.keys().next().copied()
-        {
-            warn!(
-                %peer_addr,
-                evicted = %victim,
-                cap,
-                "dead-letter pending table at cap, evicting an existing entry — \
-                 dynamic peers churning faster than they re-establish"
-            );
-            self.dead_lettered_pending.remove(&victim);
-        }
-        self.dead_lettered_pending.insert(peer_addr, entry);
-    }
-
-    /// Drain any dead-lettered hot-apply / Route Refresh intent for a
-    /// freshly accepted dynamic peer at this address and apply it to the
-    /// new `ManagedPeer`. No-op if no entry exists.
-    fn restore_dead_lettered_pending(&mut self, peer_addr: IpAddr) {
-        let Some(prev) = self.dead_lettered_pending.remove(&peer_addr) else {
-            return;
-        };
-        let Some(managed) = self.peers.get_mut(&peer_addr) else {
-            return;
-        };
-        managed.pending_refresh = prev.refresh;
-        managed.pending_export_apply = prev.export_apply;
-        managed.advertise_graceful_shutdown = prev.graceful_shutdown;
-        info!(
-            %peer_addr,
-            pending_refresh = prev.refresh,
-            pending_export_apply = prev.export_apply,
-            advertise_graceful_shutdown = prev.graceful_shutdown,
-            "restored dead-lettered hot-apply intent on dynamic peer re-establishment"
-        );
     }
 
     async fn handle_session_notification(&mut self, notification: SessionNotification) {
